comlpated

+++ src/app/laporan_admin.php
@@ -6,8 +6,16 @@ if (!isset($_SESSION['key'])) {
     exit;
 }
 
-require_once '../templates/header.php';
-require_once 'logic.php';
+require_once 'logic/logic.php';
+switch ($_COOKIE['r']) {
+    case 'guest':
+        require_once '../templates/header-guest/header.php';
+        break;
+
+    case 'resepsionis':
+        require_once '../templates/header-resepsionis/header.php';
+        break;
+}
 
 if (isset($_POST['show_entries'])) {
     // cek apakah data berhasil di tambahkan atau tidak
@@ -26,10 +34,10 @@ if (isset($_POST['show_entries'])) {
 
 <div class="container d-flex justify-content-between mt-3">
     <div class="input-group">
-        <input type="text" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default" placeholder="tanggal check out">
+        <input type="date" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default" placeholder="tanggal check out">
     </div>
     <div class="input-group ms-4 me-4">
-        <input type="text" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default" placeholder="tanggal check out">
+        <input type="date" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default" placeholder="tanggal check out">
     </div>
     <button type="button" class="btn btn-primary">search</button>
 </div>
@@ -60,8 +68,10 @@ if (isset($_POST['show_entries'])) {
                     <th scope="col">nama kamar</th>
                     <th scope="col">tanggal checkout</th>
                     <th scope="col">total biaya sewa</th>
-                    <th scope="col">edit</th>
-                    <th scope="col">delete</th>
+                    <?php if ($_COOKIE['r'] === 'resepsionis') : ?>
+                        <th scope="col">edit</th>
+                        <th scope="col">delete</th>
+                    <?php endif; ?>
                 </tr>
             </thead>
             <tbody>
@@ -84,24 +94,26 @@ if (isset($_POST['show_entries'])) {
                         <td>
                             <?= $row["total_biaya_sewa"]; ?>
                         </td>
-                        <td>
-                            <button type="button" class="btn btn-warning">
-                                <svg style="width: 16px;" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-                                    <!--! Font Awesome Pro 6.1.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. -->
-                                    <path d="M490.3 40.4C512.2 62.27 512.2 97.73 490.3 119.6L460.3 149.7L362.3 51.72L392.4 21.66C414.3-.2135 449.7-.2135 471.6 21.66L490.3 40.4zM172.4 241.7L339.7 74.34L437.7 172.3L270.3 339.6C264.2 345.8 256.7 350.4 248.4 353.2L159.6 382.8C150.1 385.6 141.5 383.4 135 376.1C128.6 370.5 126.4 361 129.2 352.4L158.8 263.6C161.6 255.3 166.2 247.8 172.4 241.7V241.7zM192 63.1C209.7 63.1 224 78.33 224 95.1C224 113.7 209.7 127.1 192 127.1H96C78.33 127.1 64 142.3 64 159.1V416C64 433.7 78.33 448 96 448H352C369.7 448 384 433.7 384 416V319.1C384 302.3 398.3 287.1 416 287.1C433.7 287.1 448 302.3 448 319.1V416C448 469 405 512 352 512H96C42.98 512 0 469 0 416V159.1C0 106.1 42.98 63.1 96 63.1H192z" />
-                                </svg>
-                                <a href="ubah.php?id=<?= $row["id_transaksi"]; ?>&param1=transaksi&param2=id_transaksi" style="text-decoration: none;">edit</a>
-                            </button>
-                        </td>
-                        <td>
-                            <button type="button" class="btn btn-danger">
-                                <svg style="width: 16px;" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-                                    <!--! Font Awesome Pro 6.1.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. -->
-                                    <path d="M490.3 40.4C512.2 62.27 512.2 97.73 490.3 119.6L460.3 149.7L362.3 51.72L392.4 21.66C414.3-.2135 449.7-.2135 471.6 21.66L490.3 40.4zM172.4 241.7L339.7 74.34L437.7 172.3L270.3 339.6C264.2 345.8 256.7 350.4 248.4 353.2L159.6 382.8C150.1 385.6 141.5 383.4 135 376.1C128.6 370.5 126.4 361 129.2 352.4L158.8 263.6C161.6 255.3 166.2 247.8 172.4 241.7V241.7zM192 63.1C209.7 63.1 224 78.33 224 95.1C224 113.7 209.7 127.1 192 127.1H96C78.33 127.1 64 142.3 64 159.1V416C64 433.7 78.33 448 96 448H352C369.7 448 384 433.7 384 416V319.1C384 302.3 398.3 287.1 416 287.1C433.7 287.1 448 302.3 448 319.1V416C448 469 405 512 352 512H96C42.98 512 0 469 0 416V159.1C0 106.1 42.98 63.1 96 63.1H192z" />
-                                </svg>
-                                <a href="hapus.php?id=<?= $row["id_transaksi"]; ?>&param1=transaksi&param2=id_transaksi" style="text-decoration: none;">hapus</a>
-                            </button>
-                        </td>
+                        <?php if ($_COOKIE['r'] === 'resepsionis') : ?>
+                            <td>
+                                <button type="button" class="btn btn-warning">
+                                    <svg style="width: 16px;" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
+                                        <!--! Font Awesome Pro 6.1.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. -->
+                                        <path d="M490.3 40.4C512.2 62.27 512.2 97.73 490.3 119.6L460.3 149.7L362.3 51.72L392.4 21.66C414.3-.2135 449.7-.2135 471.6 21.66L490.3 40.4zM172.4 241.7L339.7 74.34L437.7 172.3L270.3 339.6C264.2 345.8 256.7 350.4 248.4 353.2L159.6 382.8C150.1 385.6 141.5 383.4 135 376.1C128.6 370.5 126.4 361 129.2 352.4L158.8 263.6C161.6 255.3 166.2 247.8 172.4 241.7V241.7zM192 63.1C209.7 63.1 224 78.33 224 95.1C224 113.7 209.7 127.1 192 127.1H96C78.33 127.1 64 142.3 64 159.1V416C64 433.7 78.33 448 96 448H352C369.7 448 384 433.7 384 416V319.1C384 302.3 398.3 287.1 416 287.1C433.7 287.1 448 302.3 448 319.1V416C448 469 405 512 352 512H96C42.98 512 0 469 0 416V159.1C0 106.1 42.98 63.1 96 63.1H192z" />
+                                    </svg>
+                                    <a href="ubah.php?id=<?= $row["id_transaksi"]; ?>&param1=transaksi&param2=id_transaksi" style="text-decoration: none;">edit</a>
+                                </button>
+                            </td>
+                            <td>
+                                <button type="button" class="btn btn-danger">
+                                    <svg style="width: 16px;" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
+                                        <!--! Font Awesome Pro 6.1.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc. -->
+                                        <path d="M490.3 40.4C512.2 62.27 512.2 97.73 490.3 119.6L460.3 149.7L362.3 51.72L392.4 21.66C414.3-.2135 449.7-.2135 471.6 21.66L490.3 40.4zM172.4 241.7L339.7 74.34L437.7 172.3L270.3 339.6C264.2 345.8 256.7 350.4 248.4 353.2L159.6 382.8C150.1 385.6 141.5 383.4 135 376.1C128.6 370.5 126.4 361 129.2 352.4L158.8 263.6C161.6 255.3 166.2 247.8 172.4 241.7V241.7zM192 63.1C209.7 63.1 224 78.33 224 95.1C224 113.7 209.7 127.1 192 127.1H96C78.33 127.1 64 142.3 64 159.1V416C64 433.7 78.33 448 96 448H352C369.7 448 384 433.7 384 416V319.1C384 302.3 398.3 287.1 416 287.1C433.7 287.1 448 302.3 448 319.1V416C448 469 405 512 352 512H96C42.98 512 0 469 0 416V159.1C0 106.1 42.98 63.1 96 63.1H192z" />
+                                    </svg>
+                                    <a href="hapus.php?id=<?= $row["id_transaksi"]; ?>&param1=transaksi&param2=id_transaksi" style="text-decoration: none;">hapus</a>
+                                </button>
+                            </td>
+                        <?php endif; ?>
                     </tr>
                     <?php $i++; ?>
                 <?php endforeach; ?>
@@ -122,7 +134,8 @@ if (isset($_POST['show_entries'])) {
 
 <div class="container">
     <form action="" method="post" class="d-flex flex-row-reverse bd-highlight">
-        <button type="submit" class="btn btn-primary" id="btn_print" name="print">cetak laporan reservasi kamar hotel</button>
+        <button type="submit" class="btn btn-primary" id="btn_print" name="print">cetak laporan reservasi kamar
+            hotel</button>
         <button type="button" class="btn btn-primary me-4">cetak jurnal buku besar</button>
     </form>
 </div>
